refactor(web): replace inngest queue with workflow executor
