+++ apps/hyperlocalise-web/package.json
@@ -25,7 +25,6 @@
     "dotenv": "17.4.2",
     "drizzle-orm": "0.45.2",
     "hono": "4.12.14",
-    "inngest": "4.2.4",
     "next": "16.2.4",
     "next-themes": "0.4.6",
     "pg": "8.20.0",
@@ -41,7 +40,6 @@
     "zod": "4.3.6"
   },
   "devDependencies": {
-    "@inngest/test": "1.0.0",
     "@tailwindcss/postcss": "4.2.2",
     "@types/node": "24.12.2",
     "@types/pg": "8.20.0",
