+++ apps/hyperlocalise-web/src/api/routes/project/translation-job.route.ts
@@ -6,7 +6,7 @@ import { validator } from "hono/validator";
 
 import { type AuthVariables } from "@/api/auth/workos";
 import { db, schema } from "@/lib/database";
-import type { TranslationJobQueue } from "@/lib/inngest";
+import type { TranslationJobQueue } from "@/lib/workflow";
 
 import {
   forbiddenResponse,
