+++ apps/hyperlocalise-web/src/lib/translation/translation-job-queued-function.ts
@@ -2,7 +2,7 @@ import { and, eq, isNull, or } from "drizzle-orm";
 
 import { stringTranslationJobInputSchema } from "@/api/routes/project/translation-job.schema";
 import { db, schema } from "@/lib/database";
-import { inngest, TRANSLATION_JOB_QUEUED_EVENT } from "@/lib/inngest";
+import type { TranslationJobQueuedEventData } from "@/lib/workflow";
 import {
   translateStringJobWithOpenAI,
   type StringTranslationGenerator,
@@ -10,6 +10,11 @@ import {
 
 type CreateTranslationJobQueuedFunctionOptions = {
   translateStringJob?: StringTranslationGenerator;
+};
+
+type ExecuteTranslationJobQueuedInput = {
+  event: TranslationJobQueuedEventData;
+  runId: string;
 };
 
 function formatExecutionError(error: unknown) {
@@ -39,48 +44,31 @@ async function getStoredJob(jobId: string, projectId: string) {
   return job;
 }
 
-/**
- * Creates the workflow worker for queued translation jobs.
- *
- * Current v1 behavior:
- * - attaches the Inngest `runId` to the job record
- * - claims `queued` jobs and transitions them to `running`
- * - executes string jobs through the injected generator
- * - persists either `string_result` output or terminal error state
- *
- * File jobs are still treated as unsupported in this worker.
- */
 export function createTranslationJobQueuedFunction(
   options: CreateTranslationJobQueuedFunctionOptions = {},
 ) {
   const translateStringJob = options.translateStringJob ?? translateStringJobWithOpenAI;
 
-  return inngest.createFunction(
-    {
-      id: "translation-job-queued",
-      triggers: [{ event: TRANSLATION_JOB_QUEUED_EVENT }],
-    },
-    async ({ event, runId, step }) => {
-      const attachedJob = await step.run("attach-workflow-run-id", async () => {
-        const [job] = await db
-          .update(schema.translationJobs)
-          .set({
-            workflowRunId: runId,
-          })
-          .where(
-            and(
-              eq(schema.translationJobs.id, event.data.jobId),
-              eq(schema.translationJobs.projectId, event.data.projectId),
-              isNull(schema.translationJobs.workflowRunId),
-            ),
-          )
-          .returning({
-            id: schema.translationJobs.id,
-            projectId: schema.translationJobs.projectId,
-            type: schema.translationJobs.type,
-            runId: schema.translationJobs.workflowRunId,
-          });
-
+  return async ({ event, runId }: ExecuteTranslationJobQueuedInput) => {
+    const attachedJob = await db
+      .update(schema.translationJobs)
+      .set({
+        workflowRunId: runId,
+      })
+      .where(
+        and(
+          eq(schema.translationJobs.id, event.jobId),
+          eq(schema.translationJobs.projectId, event.projectId),
+          isNull(schema.translationJobs.workflowRunId),
+        ),
+      )
+      .returning({
+        id: schema.translationJobs.id,
+        projectId: schema.translationJobs.projectId,
+        type: schema.translationJobs.type,
+        runId: schema.translationJobs.workflowRunId,
+      })
+      .then(async ([job]) => {
         if (job) {
           return {
             ...job,
@@ -89,11 +77,11 @@ export function createTranslationJobQueuedFunction(
           };
         }
 
-        const existingJob = await getStoredJob(event.data.jobId, event.data.projectId);
+        const existingJob = await getStoredJob(event.jobId, event.projectId);
 
         if (!existingJob) {
           throw new Error(
-            `translation job ${event.data.jobId} was not found in project ${event.data.projectId}`,
+            `translation job ${event.jobId} was not found in project ${event.projectId}`,
           );
         }
 
@@ -106,202 +94,192 @@ export function createTranslationJobQueuedFunction(
         };
       });
 
-      return step.run("execute-string-job", async () => {
-        if (!attachedJob.runId) {
-          throw new Error(
-            `translation job ${event.data.jobId} does not have an associated workflow run id`,
-          );
-        }
+    if (!attachedJob.runId) {
+      throw new Error(`translation job ${event.jobId} does not have an associated workflow run id`);
+    }
 
-        if (!attachedJob.ownedByCurrentRun) {
-          return getStoredJob(event.data.jobId, event.data.projectId);
-        }
+    if (!attachedJob.ownedByCurrentRun) {
+      return getStoredJob(event.jobId, event.projectId);
+    }
 
-        const [claimedJob] = await db
-          .update(schema.translationJobs)
-          .set({
-            status: "running",
-            lastError: null,
-            outcomeKind: null,
-            outcomePayload: null,
-            completedAt: null,
-          })
-          .where(
-            and(
-              eq(schema.translationJobs.id, event.data.jobId),
-              eq(schema.translationJobs.projectId, event.data.projectId),
-              or(
-                eq(schema.translationJobs.status, "queued"),
-                eq(schema.translationJobs.status, "running"),
-              ),
-              eq(schema.translationJobs.workflowRunId, attachedJob.runId),
-            ),
-          )
-          .returning({
-            id: schema.translationJobs.id,
-            projectId: schema.translationJobs.projectId,
-            type: schema.translationJobs.type,
-            inputPayload: schema.translationJobs.inputPayload,
-          });
-
-        if (!claimedJob) {
-          const existingJob = await getStoredJob(event.data.jobId, event.data.projectId);
-
-          if (!existingJob) {
-            throw new Error(
-              `translation job ${event.data.jobId} was not found in project ${event.data.projectId}`,
-            );
-          }
-
-          return existingJob;
-        }
-
-        if (claimedJob.type !== "string") {
-          const message = `translation job type ${claimedJob.type} is not supported`;
-
-          const [failedJob] = await db
-            .update(schema.translationJobs)
-            .set({
-              status: "failed",
-              outcomeKind: "error",
-              outcomePayload: {
-                code: "unsupported_job_type",
-                message,
-              },
-              lastError: message,
-              completedAt: new Date(),
-            })
-            .where(
-              and(
-                eq(schema.translationJobs.id, claimedJob.id),
-                eq(schema.translationJobs.projectId, claimedJob.projectId),
-              ),
-            )
-            .returning({
-              id: schema.translationJobs.id,
-              projectId: schema.translationJobs.projectId,
-              type: schema.translationJobs.type,
-              status: schema.translationJobs.status,
-              workflowRunId: schema.translationJobs.workflowRunId,
-              outcomeKind: schema.translationJobs.outcomeKind,
-              outcomePayload: schema.translationJobs.outcomePayload,
-              lastError: schema.translationJobs.lastError,
-              completedAt: schema.translationJobs.completedAt,
-            });
-
-          if (!failedJob) {
-            throw new Error(
-              `translation job ${claimedJob.id} disappeared before failure could be recorded`,
-            );
-          }
-
-          return failedJob;
-        }
-
-        try {
-          const parsedInput = stringTranslationJobInputSchema.safeParse(claimedJob.inputPayload);
-
-          if (!parsedInput.success) {
-            throw new Error("invalid stored string translation job input");
-          }
-
-          const [project] = await db
-            .select({
-              name: schema.translationProjects.name,
-              translationContext: schema.translationProjects.translationContext,
-            })
-            .from(schema.translationProjects)
-            .where(eq(schema.translationProjects.id, claimedJob.projectId))
-            .limit(1);
-
-          if (!project) {
-            throw new Error(`translation project ${claimedJob.projectId} was not found`);
-          }
-
-          const result = await translateStringJob({
-            projectName: project.name,
-            projectTranslationContext: project.translationContext,
-            jobInput: parsedInput.data,
-          });
-
-          const [succeededJob] = await db
-            .update(schema.translationJobs)
-            .set({
-              status: "succeeded",
-              outcomeKind: "string_result",
-              outcomePayload: result,
-              lastError: null,
-              completedAt: new Date(),
-            })
-            .where(
-              and(
-                eq(schema.translationJobs.id, claimedJob.id),
-                eq(schema.translationJobs.projectId, claimedJob.projectId),
-              ),
-            )
-            .returning({
-              id: schema.translationJobs.id,
-              projectId: schema.translationJobs.projectId,
-              type: schema.translationJobs.type,
-              status: schema.translationJobs.status,
-              workflowRunId: schema.translationJobs.workflowRunId,
-              outcomeKind: schema.translationJobs.outcomeKind,
-              outcomePayload: schema.translationJobs.outcomePayload,
-              lastError: schema.translationJobs.lastError,
-              completedAt: schema.translationJobs.completedAt,
-            });
-
-          if (!succeededJob) {
-            throw new Error(
-              `translation job ${claimedJob.id} disappeared before success could be recorded`,
-            );
-          }
-
-          return succeededJob;
-        } catch (error) {
-          const message = formatExecutionError(error);
-
-          const [failedJob] = await db
-            .update(schema.translationJobs)
-            .set({
-              status: "failed",
-              outcomeKind: "error",
-              outcomePayload: {
-                code: "translation_execution_failed",
-                message,
-              },
-              lastError: message,
-              completedAt: new Date(),
-            })
-            .where(
-              and(
-                eq(schema.translationJobs.id, claimedJob.id),
-                eq(schema.translationJobs.projectId, claimedJob.projectId),
-              ),
-            )
-            .returning({
-              id: schema.translationJobs.id,
-              projectId: schema.translationJobs.projectId,
-              type: schema.translationJobs.type,
-              status: schema.translationJobs.status,
-              workflowRunId: schema.translationJobs.workflowRunId,
-              outcomeKind: schema.translationJobs.outcomeKind,
-              outcomePayload: schema.translationJobs.outcomePayload,
-              lastError: schema.translationJobs.lastError,
-              completedAt: schema.translationJobs.completedAt,
-            });
-
-          if (!failedJob) {
-            throw new Error(
-              `translation job ${claimedJob.id} disappeared before failure could be recorded`,
-            );
-          }
-
-          return failedJob;
-        }
+    const [claimedJob] = await db
+      .update(schema.translationJobs)
+      .set({
+        status: "running",
+        lastError: null,
+        outcomeKind: null,
+        outcomePayload: null,
+        completedAt: null,
+      })
+      .where(
+        and(
+          eq(schema.translationJobs.id, event.jobId),
+          eq(schema.translationJobs.projectId, event.projectId),
+          or(eq(schema.translationJobs.status, "queued"), eq(schema.translationJobs.status, "running")),
+          eq(schema.translationJobs.workflowRunId, attachedJob.runId),
+        ),
+      )
+      .returning({
+        id: schema.translationJobs.id,
+        projectId: schema.translationJobs.projectId,
+        type: schema.translationJobs.type,
+        inputPayload: schema.translationJobs.inputPayload,
       });
-    },
-  );
+
+    if (!claimedJob) {
+      const existingJob = await getStoredJob(event.jobId, event.projectId);
+
+      if (!existingJob) {
+        throw new Error(`translation job ${event.jobId} was not found in project ${event.projectId}`);
+      }
+
+      return existingJob;
+    }
+
+    if (claimedJob.type !== "string") {
+      const message = `translation job type ${claimedJob.type} is not supported`;
+
+      const [failedJob] = await db
+        .update(schema.translationJobs)
+        .set({
+          status: "failed",
+          outcomeKind: "error",
+          outcomePayload: {
+            code: "unsupported_job_type",
+            message,
+          },
+          lastError: message,
+          completedAt: new Date(),
+        })
+        .where(
+          and(
+            eq(schema.translationJobs.id, claimedJob.id),
+            eq(schema.translationJobs.projectId, claimedJob.projectId),
+          ),
+        )
+        .returning({
+          id: schema.translationJobs.id,
+          projectId: schema.translationJobs.projectId,
+          type: schema.translationJobs.type,
+          status: schema.translationJobs.status,
+          workflowRunId: schema.translationJobs.workflowRunId,
+          outcomeKind: schema.translationJobs.outcomeKind,
+          outcomePayload: schema.translationJobs.outcomePayload,
+          lastError: schema.translationJobs.lastError,
+          completedAt: schema.translationJobs.completedAt,
+        });
+
+      if (!failedJob) {
+        throw new Error(
+          `translation job ${claimedJob.id} disappeared before failure could be recorded`,
+        );
+      }
+
+      return failedJob;
+    }
+
+    try {
+      const parsedInput = stringTranslationJobInputSchema.safeParse(claimedJob.inputPayload);
+
+      if (!parsedInput.success) {
+        throw new Error("invalid stored string translation job input");
+      }
+
+      const [project] = await db
+        .select({
+          name: schema.translationProjects.name,
+          translationContext: schema.translationProjects.translationContext,
+        })
+        .from(schema.translationProjects)
+        .where(eq(schema.translationProjects.id, claimedJob.projectId))
+        .limit(1);
+
+      if (!project) {
+        throw new Error(`translation project ${claimedJob.projectId} was not found`);
+      }
+
+      const result = await translateStringJob({
+        projectName: project.name,
+        projectTranslationContext: project.translationContext,
+        jobInput: parsedInput.data,
+      });
+
+      const [succeededJob] = await db
+        .update(schema.translationJobs)
+        .set({
+          status: "succeeded",
+          outcomeKind: "string_result",
+          outcomePayload: result,
+          lastError: null,
+          completedAt: new Date(),
+        })
+        .where(
+          and(
+            eq(schema.translationJobs.id, claimedJob.id),
+            eq(schema.translationJobs.projectId, claimedJob.projectId),
+          ),
+        )
+        .returning({
+          id: schema.translationJobs.id,
+          projectId: schema.translationJobs.projectId,
+          type: schema.translationJobs.type,
+          status: schema.translationJobs.status,
+          workflowRunId: schema.translationJobs.workflowRunId,
+          outcomeKind: schema.translationJobs.outcomeKind,
+          outcomePayload: schema.translationJobs.outcomePayload,
+          lastError: schema.translationJobs.lastError,
+          completedAt: schema.translationJobs.completedAt,
+        });
+
+      if (!succeededJob) {
+        throw new Error(
+          `translation job ${claimedJob.id} disappeared before success could be recorded`,
+        );
+      }
+
+      return succeededJob;
+    } catch (error) {
+      const message = formatExecutionError(error);
+
+      const [failedJob] = await db
+        .update(schema.translationJobs)
+        .set({
+          status: "failed",
+          outcomeKind: "error",
+          outcomePayload: {
+            code: "translation_execution_failed",
+            message,
+          },
+          lastError: message,
+          completedAt: new Date(),
+        })
+        .where(
+          and(
+            eq(schema.translationJobs.id, claimedJob.id),
+            eq(schema.translationJobs.projectId, claimedJob.projectId),
+          ),
+        )
+        .returning({
+          id: schema.translationJobs.id,
+          projectId: schema.translationJobs.projectId,
+          type: schema.translationJobs.type,
+          status: schema.translationJobs.status,
+          workflowRunId: schema.translationJobs.workflowRunId,
+          outcomeKind: schema.translationJobs.outcomeKind,
+          outcomePayload: schema.translationJobs.outcomePayload,
+          lastError: schema.translationJobs.lastError,
+          completedAt: schema.translationJobs.completedAt,
+        });
+
+      if (!failedJob) {
+        throw new Error(
+          `translation job ${claimedJob.id} disappeared before failure could be recorded`,
+        );
+      }
+
+      return failedJob;
+    }
+  };
 }
 
-export const translationJobQueuedFunction = createTranslationJobQueuedFunction();
+export const executeTranslationJobQueued = createTranslationJobQueuedFunction();
